+++ _tasks/_consts.py
@@ -1,5 +1,6 @@
 from enum import Enum
 
 
-class Tasks(str, Enum):
+class Tasks(Enum):
     test_task = "test-task"
+    test_vault_task = "test-vault-task"