+++ gcip2/tasks_core/__init__.py
@@ -1,8 +1,10 @@
-from gcip2.tasks_core.actions import InteractivePythonAction, InteractiveShlex
+from gcip2.tasks_core.models.actions import InteractivePythonAction, InteractiveShlex
+from gcip2.tasks_core.models.params import Param
 from gcip2.tasks_core.task_builder import TaskBuilderImpl
 
 __all__ = (
     "InteractivePythonAction",
     "InteractiveShlex",
     "TaskBuilderImpl",
+    "Param",
 )