+++ gcip2/tasks_core/paramsbuilder/__init__.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-
-class Param(BaseModel):
-    pass