+++ tests/unit/_test_tasks/__init__.py
@@ -1,0 +1,3 @@
+from ._tasks import TASK_GENERATOR
+
+__all__ = ("TASK_GENERATOR",)